[DEV] Implementing the GetProjectUseCase
WIthin Clean Architecture, the Domain layer defines Use Cases which define operations which can be performed in our application. In this lesson, we’ll be implementing the Get Projects Use Cases classes for our application so that we can retrieve a list of projects from the external data source through the domain layer of our Android project.

In Clean Architecture, the Domain layer defines Use Cases which define operations which can be performed in our application. In this lesson, we’ll be implementing the Get Bookmarked Projects Use Cases class for our application so that we can retrieve a list of bookmarked projects within our external data source through our domain layer.