[DEV] map the Project from Data-layer to Domain-layer
The Data layer allows us to abstract the sources of the data that our application uses. In this lesson, we’re going to begin building our data layer by creating the data models for use in its operations, as well as creating the mapper classed used to map between this representation and the Domain layers representation of the project class.