[DEV] Creating the Retrofit Service Factory for the remote service interface
The Remote layer allows us to abstract the remote source of data that our application uses. To make the requests to the external source we're making use of the Retrofit library. In this lesson, we’re going to create the Retrofit Service Factory which will provide our retrofit instance for us to make API requests in our application.