[DEV] Creating the data model mapper for the Remote Layer
Within Clean Architecture, the Remote layer allows us to abstract the remote source of data that our application uses. In this lesson, we're going to be creating the mapper classes used to map the data we get back from the API request into a format that is compatible with inner layers.