[DEV] Creating the Remote Source implementation
Within Clean Architecture, the Remote layer allows us to abstract the remote source of data that our application uses. In this lesson, we're going to be creating the class which will be retrieving data from our external remote data source, which is the implementation of the remote interface from the Data module.