[DEV] Creating the data model mapper for the presentation layer
The Presentation layer allows us to abstract the presentation logic from the framework specific implementation of our user interface. In this lesson, we're going to be creating the data model mapper for this layer so that we can map our domain layer model representations to a format that belongs to this layer.

Co-Authored-By: krongnang <krongnang@users.noreply.github.com>
Co-Authored-By: Nam Pham <nam2210@users.noreply.github.com>