[DEV] Creating the presentation state model
The Presentation layer allows us to abstract the presentation logic from the framework specific implementation of our user interface. In this lesson, we'll be recreating a ResourceState and Resource class so that we can provide presentation data to layers outside of the presentation layer in a consistent and abstract manner.

Co-Authored-By: krongnang <krongnang@users.noreply.github.com>
Co-Authored-By: Nam Pham <nam2210@users.noreply.github.com>