[DEV] Creating the Post Execution Thread Implementation
The UI module allows us to implement the user interface of our application for a specific form factor or framework. In this lesson, we're going to be creating an implementation of the PostExecutionThread interface that has allowed us to provide an abstraction for other layers for when it comes to the threading method used for work to be executed on.