[DEV] Setting up the application class
The UI module allows us to implement the user interface of our application for a specific form factor or framework. In this lesson, we're going to be implementing the application class that will be launched by the system when our application is started, allowing us to perform any necessary configuration during launch.